requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.25.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
PyYAML>=6.0
tqdm>=4.65.0
numpy>=1.24.0
